feat: Add TaskList component and implement delete functionality
The code changes add the TaskList component to the Main.js file and implement the delete functionality in the TaskList.js file. The TaskList component displays a list of tasks fetched from the server, and each task has a delete button that triggers a request to delete the task from the server. This change was made in response to the recent user commit "Client: added task form".